+++ tests/data_parsers/event_data_parsers/test_metrica_event_data_parser.py
@@ -24,17 +24,6 @@ class TestMetricaEventDataParser(unittest.TestCase):
         self.md_loc = "tests/test_data/metrica_metadata_test.xml"
         self.ed_loc = "tests/test_data/metrica_event_data_test.json"
 
-    def test_get_event_data(self):
-        expected_event_data = ED_METRICA.copy()
-        for col in ["end_x", "start_x"]:
-            expected_event_data[col] = (expected_event_data[col] + 50) / 100.0
-        for col in ["end_y", "start_y"]:
-            expected_event_data[col] = (expected_event_data[col] + 25) / 50.0
-
-        expected_event_data.drop(["datetime"], axis=1, inplace=True)
-        ed = _get_event_data(self.ed_loc)
-        pd.testing.assert_frame_equal(ed, expected_event_data)
-
     def test_load_metrica_event_data(self):
         ed, md, dbpe = load_metrica_event_data(self.ed_loc, self.md_loc)
         assert md == MD_METRICA_ED
@@ -45,6 +34,23 @@ class TestMetricaEventDataParser(unittest.TestCase):
 
         with self.assertRaises(TypeError):
             load_metrica_event_data(22, self.md_loc)
+
+    def test_load_metrica_event_data_errors(self):
+        with self.assertRaises(FileNotFoundError):
+            load_metrica_event_data(self.ed_loc[3:], self.md_loc)
+        with self.assertRaises(FileNotFoundError):
+            load_metrica_event_data(self.ed_loc, self.md_loc[1:])
+
+    def test_get_event_data(self):
+        expected_event_data = ED_METRICA.copy()
+        for col in ["end_x", "start_x"]:
+            expected_event_data[col] = (expected_event_data[col] + 50) / 100.0
+        for col in ["end_y", "start_y"]:
+            expected_event_data[col] = (expected_event_data[col] + 25) / 50.0
+
+        expected_event_data.drop(["datetime"], axis=1, inplace=True)
+        ed = _get_event_data(self.ed_loc)
+        pd.testing.assert_frame_equal(ed, expected_event_data)
 
     @patch(
         "requests.get",