+++ tests/test_visualize.py
@@ -1,10 +1,14 @@
 import os
 import unittest
+from unittest.mock import patch
 
 import matplotlib.pyplot as plt
+import numpy as np
 import pandas as pd
 
 from databallpy.get_match import get_match
+from databallpy.utils.errors import DataBallPyError
+from databallpy.utils.warnings import DataBallPyWarning
 from databallpy.visualize import (
     plot_events,
     plot_soccer_pitch,
@@ -44,6 +48,9 @@ class TestVisualize(unittest.TestCase):
         self.assertIsInstance(ax, plt.Axes)
         self.assertEqual(len(ax.lines), 27)
         self.assertEqual(len(ax.collections), 3)
+
+        with self.assertRaises(ValueError):
+            plot_soccer_pitch(pitch_color="invalid_color")
 
     def test_plot_events(self):
         match = self.match.copy()
@@ -116,6 +123,18 @@ class TestVisualize(unittest.TestCase):
         self.assertEqual(ax.get_title(), "My Test Plot3")
         self.assertEqual(len(ax.collections), 4)
 
+    def test_plot_events_wrong_input(self):
+        with self.assertRaises(TypeError):
+            plot_events(self.match.copy(), events="passes")
+        with self.assertRaises(ValueError):
+            plot_events(self.match.copy(), events=["tackle"])
+        with self.assertRaises(ValueError):
+            plot_events(self.match.copy(), outcome="false")
+        with self.assertRaises(ValueError):
+            plot_events(self.match.copy(), player_ids=[1, 2])
+        with self.assertRaises(ValueError):
+            plot_events(self.match.copy(), team_id="12345678")
+
     def test_save_match_clip(self):
         match = self.match.copy()
         match.tracking_data["player_possession"] = [
@@ -126,7 +145,7 @@ class TestVisualize(unittest.TestCase):
             "away_17",
         ]
         series = pd.Series([22, 23, 25], index=[1, 2, 3])
-        with self.assertRaises(AssertionError):
+        with self.assertRaises(DataBallPyError):
             save_match_clip(
                 match,
                 1,
@@ -135,7 +154,7 @@ class TestVisualize(unittest.TestCase):
                 title="test_clip",
                 events=["pass"],
             )
-        with self.assertRaises(AssertionError):
+        with self.assertRaises(DataBallPyError):
             save_match_clip(
                 match,
                 0,
@@ -144,7 +163,7 @@ class TestVisualize(unittest.TestCase):
                 title="test_clip",
                 variable_of_interest=series,
             )
-        with self.assertRaises(AssertionError):
+        with self.assertRaises(DataBallPyError):
             save_match_clip(
                 match,
                 1,
@@ -154,8 +173,18 @@ class TestVisualize(unittest.TestCase):
                 variable_of_interest=series,
                 player_possession_column="unknown_column",
             )
+        with self.assertRaises(DataBallPyError):
+            save_match_clip(
+                match,
+                1,
+                3,
+                save_folder="tests/test_data",
+                title="test_clip",
+                add_pitch_control=True,
+            )
 
-        assert not os.path.exists("tests/test_data/test_clip.mp4")
+        if os.path.exists("tests/test_data/test_clip.mp4"):
+            os.remove("tests/test_data/test_clip.mp4")
 
         save_match_clip(
             match,
@@ -170,7 +199,8 @@ class TestVisualize(unittest.TestCase):
         assert os.path.exists("tests/test_data/test_clip.mp4")
         os.remove("tests/test_data/test_clip.mp4")
 
-    def test_save_match_clip_with_events(self):
+    @patch("databallpy.visualize.get_pitch_control_period")
+    def test_save_match_clip_with_events(self, mock_get_pitch_control_period):
         synced_match = get_match(
             tracking_data_loc="tests/test_data/sync/tracab_td_sync_test.dat",
             tracking_metadata_loc="tests/test_data/sync/tracab_metadata_sync_test.xml",
@@ -180,7 +210,31 @@ class TestVisualize(unittest.TestCase):
             event_data_provider="opta",
             check_quality=False,
         )
+        mock_get_pitch_control_period.return_value = np.zeros(
+            (len(synced_match.tracking_data), 240, 360)
+        )
+        for col in (
+            synced_match.home_players_column_ids()
+            + synced_match.away_players_column_ids()
+        ):
+            mask = synced_match.tracking_data[col + "_x"].notnull()
+            synced_match.tracking_data.loc[mask, col + "_vx"] = 2
+            synced_match.tracking_data.loc[mask, col + "_vy"] = 2
+            synced_match.tracking_data.loc[mask, col + "_velocity"] = 2.67
         synced_match.allow_synchronise_tracking_and_event_data = True
+
+        grid_size = [360, 240]
+        x_range = np.linspace(
+            -100 / 2 - 5,
+            100 / 2 + 5,
+            grid_size[0],
+        )
+        y_range = np.linspace(
+            -50 / 2 - 5,
+            50 / 2 + 5,
+            grid_size[1],
+        )
+        grid = np.meshgrid(x_range, y_range)
 
         synced_match.synchronise_tracking_and_event_data(n_batches=2)
         events = [
@@ -189,16 +243,25 @@ class TestVisualize(unittest.TestCase):
             "shot",
         ]
 
-        assert not os.path.exists("tests/test_data/test_match_with_events.mp4")
+        if os.path.exists("tests/test_data/test_match_with_events.mp4"):
+            os.remove("tests/test_data/test_match_with_events.mp4")
 
-        save_match_clip(
-            synced_match,
-            1,
-            10,
-            save_folder="tests/test_data",
-            title="test_match_with_events",
-            events=events,
-        )
+        with self.assertWarns(DataBallPyWarning):
+            save_match_clip(
+                synced_match,
+                1,
+                10,
+                save_folder="tests/test_data",
+                title="test_match_with_events",
+                events=events,
+                add_pitch_control=True,
+                add_velocities=True,
+                verbose=False,
+            )
 
         assert os.path.exists("tests/test_data/test_match_with_events.mp4")
         os.remove("tests/test_data/test_match_with_events.mp4")
+        grid_res = mock_get_pitch_control_period.call_args.kwargs["grid"]
+        td_res = mock_get_pitch_control_period.call_args.kwargs["tracking_data"]
+        pd.testing.assert_frame_equal(synced_match.tracking_data.loc[1:10], td_res)
+        np.testing.assert_allclose(grid_res, grid)